yak. fix build on windoze.
