+++ server.go
@@ -5,16 +5,12 @@ import (
 	"crypto/rand"
 	"crypto/subtle"
 	"encoding/binary"
-	"fmt"
 	"io"
 	"log"
 	"net"
-	"os"
-	"os/signal"
 	"runtime"
 	"sync"
 	"sync/atomic"
-	"syscall"
 	"time"
 
 	"golang.org/x/crypto/ed25519"
@@ -287,27 +283,4 @@ func RunServer(conf Conf) {
 		conns <-conn
 	}
 
-}
-
-func handleSignals() {
-        signals := make(chan os.Signal, 1)
-        signal.Notify(signals, syscall.SIGINT, syscall.SIGQUIT, syscall.SIGTERM, syscall.SIGUSR1)
-
-	for {
-		select {
-		case signal, ok := <-signals:
-			if !ok { break }
-
-			switch(signal) {
-			case syscall.SIGINT, syscall.SIGQUIT, syscall.SIGTERM:
-				os.Exit(2)
-			case syscall.SIGUSR1:
-				if len(storedContent.ciphertextWithNonce) > 0 {
-					fmt.Printf("%v: some data is stored\n", os.Args[0])
-				} else {
-					fmt.Printf("%v: no data stored yet\n", os.Args[0])
-				}
-			}
-		}
-	}
 }
